requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
isort>=5.13.0
ruff>=0.1.0

